Replace delimiters
Objectives: Change the desired delimiter used in a string or a single column.
Solution: Data is using an unexpected delimiter (e.g., '%[ -()#.,=<>]/\|%') and needs to be standardized to a certain instance, such as '_' or '~' (interchangeably).

Noted: Delimiter is a single character. For multi-character delimiters, a different approach may be needed.